+++ src/re_com/alert.cljs
@@ -110,7 +110,8 @@
                :post-props {:src (at)}
                :props      {:children
                             [(part ::ab/body-wrapper
-                               {:props {:children [(part ::ab/body {})]}})
+                               {:props {:children [(part ::ab/body
+                                                     {:theme theme})]}})
                              (when (and (not heading-provided?) closeable? on-close)
                                close-alert)]}}))]}}))))
 
@@ -150,7 +151,6 @@
      {:name :parts        :required false                                :type "map"                     :validate-fn (parts? alert-list-parts) :description "See Parts section below."}
      {:name :src          :required false                                :type "map"                     :validate-fn map?                      :description [:span "Used in dev builds to assist with debugging. Source code coordinates map containing keys" [:code ":file"] "and" [:code ":line"]  ". See 'Debugging'."]}
      {:name :debug-as     :required false                                :type "map"                     :validate-fn map?                      :description [:span "Used in dev builds to assist with debugging, when one component is used implement another component, and we want the implementation component to masquerade as the original component in debug output, such as component stacks. A map optionally containing keys" [:code ":component"] "and" [:code ":args"] "."]}]))
-
 
 (defn alert-list
   "Displays a list of alert-box components in a v-box. Sample alerts object: